add tool\Output\Extra, consistent admin message if extra content is missing
